requirements.txt

kgraphservice>=0.0.7
requests
uvicorn[standard]==0.27.0.post1
fastapi==0.109.2
pydantic==2.9.1
googlemaps==4.10.0
pyyaml
google-search-results>=2.4.2
vital-llm-cluster-mgr>=0.0.5
httpx
vital-agentbox>=0.0.2
PyJWT>=2.8.0
python-dotenv
mailgun
